// types/authType.ts
export type SignIn = {
    phone: string;
    password: string;
  };
  
  export type Register = {
    username: string;
    email: string;
    password: string;
    phone: string;
    roleId: 3;
  };
  
  
  export type SuplierFields = {
    supplierName: string;
    supplierSurname: string;
    phone: string;
   
  };
  
  